[fix] Extracted GenerateReceiptPdf methods to service class - deleted utils class
